Use packages instead of being a submodule of opendylan
Having the website repo be a submodule of opendylan and linking into its
documentation directory was kind of a quick fix when opendylan.org was being
revamped to use Furo theme and integrate the web site and the documentation
into a single index. It was problematic because it meant updating the website
submodule every time something changed, which is cumbersome and easy to forget.

With this change, the web site can deploy from the master branch and the
documentation bundled with Open Dylan will update whenever a new version of
opendylan is published in the pacman catalog.

A bug in update.sh was fixed: files are copied to $dest_dir instead of to
/var/www/opendylan.org.

The site generated by this change is live on opendylan.org. There are a couple
of warnings that haven't been addressed yet, but they're not serious, and
update-opendylan.org.sh has not yet been tested or installed on opendylan.org.